+++ .babelrc
@@ -1,0 +1,9 @@
+{
+    "plugins": [
+        ["transform-runtime",
+        {
+            "polyfill": false,
+            "regenerator": true
+        }]
+    ]
+}
