+++ src/js/auth-success.js
